requirements.txt
streamlit==1.32.0
plotly==5.18.0
pandas==2.2.0
openpyxl==3.1.2
fpdf2==2.7.8
python-dotenv==1.0.1
facebook-business==19.0.0
google-ads==23.1.0
numpy==1.26.4
pillow==10.2.0
requests==2.31.0 